add - persistent storage with redux-persist and async storage
This commit adds persistent storage to the app
using redux-persist and async storage.

The redux-persist library automatically saves the
state of the app to the device's storage and retrieves
it when the app is opened again. We don't need to
save and load the async storage data manually.

Changes to be committed:
	modified:   App.js
	modified:   package-lock.json
	modified:   package.json
	modified:   store.js